+++ packages/basehub/src/bin/main.ts
@@ -4,6 +4,7 @@ import path from "path";
 import { Args } from ".";
 import fs from "fs";
 import * as esbuild from "esbuild";
+import { ScssModulesPlugin } from "esbuild-scss-modules-plugin";
 import {
   getStuffFromEnv,
   runtime__getStuffFromEnvString,
@@ -117,7 +118,11 @@ export const main = async (
 
     // this should go at the end so that it doesn't suffer any modifications.
     schemaFileContents = schemaFileContents.concat(
-      `\n${runtime__getStuffFromEnvString({ ...options, draft })}}`
+      `\n${runtime__getStuffFromEnvString({
+        ...options,
+        draft,
+        forceDraft: opts?.forceDraft,
+      })}}`
     );
 
     // 3. append our basehub function to the end of the file.
@@ -143,10 +148,12 @@ export const main = async (
       "react-dom",
       "../index",
       "@basehub/mutation-api-helpers",
+      "next",
     ];
 
     logIfNotSilent(silent, "📦 Compiling to JavaScript...");
     const reactPumpOutDir = path.join(basehubOutputPath, "react-pump");
+    const nextToolbarOutDir = path.join(basehubOutputPath, "next-toolbar");
     await Promise.all([
       esbuild.build({
         entryPoints: [generatedMainExportPath],
@@ -192,6 +199,41 @@ export const main = async (
           },
         ],
       }),
+      esbuild.build({
+        entryPoints: [
+          path.join(basehubModulePath, "src", "next", "toolbar", "index.ts"),
+        ],
+        bundle: true,
+        outdir: nextToolbarOutDir,
+        minify: false,
+        treeShaking: true,
+        splitting: true,
+        format: "esm",
+        target: ["es2020", "node18"],
+        external: peerDependencies,
+        plugins: [
+          ScssModulesPlugin(),
+          {
+            name: "use-client-for-client-components",
+            setup(build) {
+              build.onEnd(() => {
+                const rxp = /['"]use client['"]\s?;/i;
+                const outputFilePaths = fs.readdirSync(nextToolbarOutDir);
+                outputFilePaths
+                  ?.filter((fileName) => !fileName.endsWith(".map"))
+                  .forEach((fileName) => {
+                    // if the file contains "use client" we'll make sure it's on the top.
+                    const filePath = path.join(nextToolbarOutDir, fileName);
+                    const fileContents = fs.readFileSync(filePath, "utf-8");
+                    if (!rxp.test(fileContents)) return;
+                    const newContents = fileContents.replace(rxp, "");
+                    fs.writeFileSync(filePath, `"use client";\n${newContents}`);
+                  });
+              });
+            },
+          },
+        ],
+      }),
     ]);
 
     appendGeneratedCodeBanner(basehubOutputPath, args["--banner"]);
@@ -223,6 +265,14 @@ export const main = async (
         basehubModulePath,
         "react-pump.d.ts"
       );
+      const nextToolbarIndexJsPath = path.join(
+        basehubModulePath,
+        "next-toolbar.js"
+      );
+      const nextToolbarIndexDtsPath = path.join(
+        basehubModulePath,
+        "next-toolbar.d.ts"
+      );
       fs.writeFileSync(
         indexJsPath,
         `module.exports = require("${path.relative(
@@ -249,6 +299,20 @@ export const main = async (
         `export * from "${path.relative(
           basehubModulePath,
           path.join(reactPumpOutDir, "index.d.ts")
+        )}";`
+      );
+      fs.writeFileSync(
+        nextToolbarIndexJsPath,
+        `module.exports = require("${path.relative(
+          basehubModulePath,
+          path.join(nextToolbarOutDir, "index.js")
+        )}");`
+      );
+      fs.writeFileSync(
+        nextToolbarIndexDtsPath,
+        `export * from "${path.relative(
+          basehubModulePath,
+          path.join(nextToolbarOutDir, "index.d.ts")
         )}";`
       );
     }