+++ packages/basehub/src/next/toolbar/index.ts
@@ -1,0 +1,1 @@
+export { ServerToolbar as Toolbar } from "./server-toolbar";